feat(roadmaps): refactor roadmap generation and roadmap UI experience
- Refactor roadmap graph into smaller components, hooks, types, constants, and utilities
- Add stack list, skill tree, display mode switch, search input, filter drawer, legend components
- Split React Flow node, edge, marker, handle, and layout logic into focused modules
- Improve roadmap filtering flow with URL-backed status, node type, and search filters
- Add reusable graph loading, empty, and error states with retry handling
- Update stack list rendering with grouped, milestone, orphan, child-node, badge progress icon logic
- Calculate in-progress group icons from completed child-node percentage thresholds
- Suppress root hydration warnings caused by extension-injected html/body attributes
- Improve API roadmap generation validation, normalization fallback, and loggin
