+++ apps/web/app/(full-layout)/roadmaps/[id]/_hooks/use-responsive-drawer-direction.ts
@@ -1,0 +1,21 @@
+'use client';
+
+import { useEffect, useState } from 'react';
+
+export function useResponsiveDrawerDirection() {
+  const [drawerDirection, setDrawerDirection] = useState<'bottom' | 'right'>('bottom');
+
+  useEffect(() => {
+    const mediaQuery = window.matchMedia('(min-width: 1024px)');
+    const updateDrawerDirection = () => {
+      setDrawerDirection(mediaQuery.matches ? 'right' : 'bottom');
+    };
+
+    updateDrawerDirection();
+    mediaQuery.addEventListener('change', updateDrawerDirection);
+
+    return () => mediaQuery.removeEventListener('change', updateDrawerDirection);
+  }, []);
+
+  return drawerDirection;
+}
