+++ apps/web/app/(full-layout)/roadmaps/[id]/_types/roadmap-display-mode.types.ts
@@ -1,0 +1,1 @@
+export type RoadmapDisplayMode = 'stack-list' | 'skill-tree';
