+++ apps/web/app/(full-layout)/roadmaps/[id]/_types/roadmap-filter.types.ts
@@ -1,0 +1,7 @@
+import type { NodeType, ProgressStatus } from './roadmap-node.types';
+
+export interface UpdateRoadmapFiltersOptions {
+  nodeType?: NodeType | null;
+  q?: string | null;
+  status?: ProgressStatus | null;
+}
